requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
openpyxl>=3.1.0
gspread>=6.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
